[TASK] Revert renderable concept
Patch 70046ab6 introduced a "concept of Renderables".

The patch claimed:
"The feature is also the basis for a larger refactoring
which finally stop storing sections in the
VariableProvider while parsing, by providing a proper
API for getting sections from a parsed template object.".

This never happened. In the end, "Renderables" is a
completely unused codebase without any usages in the
wild, that has not been followed upon to put it into
action.

It does not seem to help us much with the issue that
layout and sections are resolved at compile time and
stored in variable provider - the issue still exists
and a solution will not use this concept.

The patch essentially reverts the construct introduced
with 70046ab6 again.

This is not considered breaking since the construct
had no practical use and people never used it.