[TASK] Raise phpunit and align changes
PHPUnit removed the `-v` cli argument
with 10.2.4 as a followup. That flag
was the short form for the --verbose
removed with 10.0 and did not had any
effect.

Due to the fact that no composer.lock
file is used, the newer version is
installed leading to failing tests.

This change raises the version of
PHPUnit and aligns the test calls
to avoid the removed cli argument.

Used command(s):

```shell
composer require --dev --no-update \
    "phpunit/phpunit":"^10.2.6"
```