[BUGFIX] Throw array conversion exception in `AbstractNode->castToString()`
`AbstractNode->castToString()` already checked for objects which could not
converted to a string representation due to missing `__toString()` method.
It's possible that the value could be an array, which per nature is not
really convertable to a string anyway. Older php versions simply converted
it to an `Array` string along with a notices, which has been raised to a
warning in newer PHP versions.

There is no way to properly convert an array, therefore the PHP warning
should always lead to an parser exception like for object not convertable.

This change now checks explicitly for an array and throws a corresponding
`TYPO3Fluid\Fluid\Core\Parser\Exception` with a meaningfull message.

Test is extended to cover this case.

Resolves: #826