[TASK] Remove constructor from variable provider interface
Currently, the VariableProviderInterface suggests that its implementations
can receive an array of variables as a constructor argument to be used
as initial state of the provider instance. However, this assumption is
already not correct anymore because ChainedVariableProvider receives
an array of VariableProvider instances instead. In addition, this
restricts possible other implementations of the interface.

Noteworthy in the code and relevant projects using Fluid:

* Neos implements its own VariableProvider which extends StandardVariableProvider. The change has no effect there.
https://github.com/neos/fluidadaptor/blob/8.3/Classes/Core/ViewHelper/TemplateVariableContainer.php

* Fluid uses the constructor mostly with concrete implementations (mostly StandardVariableProvider, but also ChainedVariableProvider).

* ConsoleRunner uses a dynamic class name, but no constructor arguments.
https://github.com/TYPO3/Fluid/blob/main/src/Tools/ConsoleRunner.php#L116C68-L116C68

* flux, vhs and fluid-parameters only use StandardVariableProvider

* Fluid Components and Fluid Styleguide only use StandardVariableProvider