[FEATURE] Proper support of local variables in ViewHelpers (#828)
In its current implementation, the `f:alias` ViewHelper has side effects:
It adds variables to the global variable container, which are not restored
properly after the ViewHelper call. If a variable name is used as alias
which already exists, that variable is unset after the ViewHelper call.

To resolve this, a new `ScopedVariableProvider` is introduced to handle
local variables inside ViewHelper calls properly. After the local scope,
global variables are restored to their original value in case they existed.
At the same time, other modifications to the variable provider inside
the `f:alias` call, such as a call to `f:variable`, should still be persistent
after the ViewHelper call. This ensures that the change is non-breaking
for the vast majority of usages. The only situation where this could be
breaking is when you relied on the arguably broken behavior that `f:alias`
unsets the variables afterwards.

This is also the reason why we can't simple store and reset the whole
variable container. This would be a breaking change and would also be
a diversion from other ViewHelpers, such as `f:if`, where `f:variable` also
leaks out. To change this behavior for select ViewHelpers would be
inconsistent and hard to understand for users of Fluid.

before:

```xml
<f:variable name="myVariable" value="initialValue" />
<f:alias map="{myVariable: \'myValue\'}">...</f:alias>
{myVariable} <!-- outputs empty string -->
```

after:

```xml
<f:variable name="myVariable" value="initialValue" />
<f:alias map="{myVariable: \'myValue\'}">...</f:alias>
{myVariable} <!-- outputs "initialValue" -->
```

This behavior also applies to other ViewHelpers that provide variables
to their child nodes, such as `f:for`, `f:render`, `f:groupedFor`, `f:cycle`.
These will be modified accordingly in separate PRs.

Resolves #666