[FEATURE] Attributes for RenderingContext (#832)
This adds the ability to attach arbitrary attributes to the rendering
context. All popular usages of Fluid out there extend the rendering
context implementation of Fluid and don't actually implement the
whole interface, so we should be able to change the interface as
well.

As a result of this, TYPO3 CMS will be able to drop its own rendering
context implementation, which will make the integration more
straightforward.