[BUGFIX] Fix memory overhead of getScopedCopy()
The previous implementation of getScopedCopy() in the new
ScopedVariableProvider copied both local and global variables.
This is not necessary because all variables (except for settings)
are provided explicitly and thus existing data in the data
structure is irrelevant.

The new implementation reduces that unnecessary overhead
by returning a StandardVariableProvider instead.