[BUGFIX] Explicitly set variables override local variables
The default behavior of the newly introduced local and global
variable scopes has been that existing local variables are
preferred over global variables if both exist. While this makes
sense in general, if a local variable gets overwritten by a
VariableViewHelper, this value couldn't be used in the local
context because the local variable still had its old value.

This change makes sure that both the local and the global
variable get the new value. This eliminates the described
inconsistencies when using a name of a local variable for
a global variable.