[FEATURE] Allow comparison with null in `<f:if>` ViewHelper
It is not possible to use a condition with the `<f:if />`
ViewHelper to check directly against null value, like it
is known from the PHP world.

For example, following php-code

```php
  // strict comparison
  if ($variable === null) {}
  if ($variable !== null) {}

  // loose comparision
  if ($variable == null) {}
  if ($variable != null) {}
```

can now be written with fluid as

```xhtml
  <!-- /* strict comparison */ -->
  <f:if condition="{var} === null"></f:if>
  <f:if condition="{var} !== null"></f:if>

  <!-- /* loose comparison */ -->
  <f:if condition="{var} == null"></f:if>
  <f:if condition="{var} != null"></f:if>
```

which extinguish the need of a custom `isNull` ViewHelper
in projects.