[BUGFIX] Prevent php warning when passing array as tag attribute (#858)
TagBuilder::addAttribute() allows both strings and array/traversables
as attribute value. However, in the implementation only "data" and
"aria" are properly handled as arrays. For other attribute names,
this leads to a type conversion warning.

With this patch, that edge case is properly handled with an
exception. Also, test coverage is added both for data/aria and the
exception.

Resolves: #857