[FEATURE] JoinViewHelper
The JoinViewHelper combines elements from an array into a single string. You can specify both a general separator and a special one for the last element that is rendered between the elements.

```html
<f:join value="{0: '1', 1: '2', 2: '3'}" />
```

Results in the output:

```html
123
```

```html
<f:join value="{0: '1', 1: '2', 2: '3'}" separator=", " />
```

Results in the output:

```html
1, 2, 3
```

```html
<f:join value="{0: '1', 1: '2', 2: '3'}" separator=", " separatorLast=" and " />
```

Results in the output:

```html
1, 2 and 3
```