[TASK] Deprecate shortcut methods in AbstractTemplateView
In order to simplify integrations of the Fluid template engine into
frameworks (such as TYPO3), we deprecate four methods of
AbstractTemplateView that are only shortcuts to the
rendering context object.

These methods will be removed in Fluid v4 because the migration
path is easy and it simplifies further work on TYPO3's Fluid
integration.