[!!!][TASK] Add type hints to template compiler (#948)
This patch adds type hints to the template compiler and associated
classes, such as parsed/compiled templates and the parsing state.
A few adjustments in test cases are necessary because of
excessive mocking of internal functionality. In some cases, types
were used in mock implementations that don't actually exist in
real world use, these are now adjusted when necessary.

Note that these changes require all cached templates to be
regenerated because type hints need to be added to the generated
classes as well. Because of this, the patch has been marked as
breaking.