[!!!][FEATURE] Boolean and Null literals
This patch enhances Fluid's capabilities of dealing with booleans
and null significantly. Previously, the actual booleans `false` and `true`
could be only created fron within inline syntax or if a ViewHelper
argument was defined as boolean (in which case strings were parsed
as boolean). `null` values couldn't really be created at all. Now, there
exist standard variables for true, false, and null, which can be used in
all contexts.

This patch is marked as breaking because it disallows the usage of
the variables `true`, `false` and `null` in variable providers.