[BUGFIX] Allow leading/trailing whitespace in math expression
The parser of MathExpressionNode already allowed whitespace before
or after an expression. However, ExpressionNodes are only executed
if their `$detectionExpression` evaluates to true. This was not the case
for leading/trailing whitespace, which is why the regular expression is
extended to allow this.

Unfortunately, the test did only evaluate the parser but did not check
the regular expression beforehand. As there is no API for this (the
regular expression is the API and is executed directly in
`TemplateParser::textAndShorthandSyntaxHandler()`, we execute
the regex manually in the test as well. One existing test case is
removed because it never was actually supported.

In the future, we probably should implement a proper API for this.