[FEATURE] Fluid code in comments no longer interpreted
Previously, the behavior of the `<f:comment>` ViewHelper was confusing:
Even if invalid Fluid code (e. g. syntax errors or a non-existent ViewHelper)
was wrapped with `<f:comment>`, it would still be interpreted by the
Fluid parser and an exception would be thrown.

This patch resolves this similarly to CDATA handling. It introduces a new
TemplateProcessor, which is executed before any other parsing and strips
out the contents of Fluid comments. The `<f:comment>` ViewHelpers are
kept intact and are filled with empty lines. This can be helpful when
debugging Fluid templates because the line numbers in exceptions are
consistent to the original template file.

This new TemplateProcessor is then registered in the default
RenderingContext implementation. For systems that set their own
TemplateProcessors, this would probably need adjustments. However,
this change is not breaking since the old functionality is still there.

Fluid's own tests however now assume that this new TemplateProcessor
is registered and are adjusted properly in this patch.