[BUGFIX] Respect getContentArgumentName() in TemplateParser
With the introduction of the new `getContentArgumentName()`
API for ViewHelpers, there has been an oversight in `TemplateParser`:
It had special handling for the old method name
`resolveContentArgumentName()`, which is now deprecated. Since
`resolveContentArgumentName()` is now called by the new API
`getContentArgumentName()`, we can just call the new API in the
`TemplateParser` now, without special handling for the deprecated
trait.

As a consequence, we need to adjust some tests and fixtures. Also,
we add more dedicated test cases to our new test file.

In the future, there needs to be a cleanup of `AbstractViewHelper`
and `ViewHelperInterface`, since internal classes depend on API
only defined in `AbstractViewHelper`. In fact, we arguably made things
a bit worse now, but this is a topic for Fluid v5.